# Optimization for batched processing