--- a/data_juicer/ops/base_op.py
+++ b/data_juicer/ops/base_op.py
@@ -289,6 +289,20 @@ class Filter(OP):
         else:
             self.compute_stats = catch_map_single_exception(self.compute_stats)
 
+    def compute_stats_batched(self, samples, **kwargs):
+        keys = samples.keys()
+        samples_stats = samples[Fields.stats]
+        for i, stat in enumerate(samples_stats):
+            this_sample = {key: samples[key][i] for key in keys}
+            res_sample = self.compute_stats(this_sample, **kwargs)
+            samples[Fields.stats][i] = res_sample[Fields.stats]
+
+        return samples
+
+    def process_batched(self, samples):
+        return map(lambda stat: self.process({Fields.stats: stat}),
+                   samples[Fields.stats])
+
     def compute_stats(self, sample, context=False):
         """
         Compute stats for the sample which is used as a metric to decide
@@ -322,14 +336,14 @@ class Filter(OP):
                                   num_proc=self.runtime_np(),
                                   batch_size=self.batch_size,
                                   desc='Adding new column for stats')
-        dataset = dataset.map(self.compute_stats,
+        dataset = dataset.map(self.compute_stats_batched,
                               num_proc=self.runtime_np(),
                               with_rank=self.use_cuda(),
                               batch_size=self.batch_size,
                               desc=self._name + '_compute_stats')
         if exporter and self.stats_export_path is not None:
             exporter.export_compute_stats(dataset, self.stats_export_path)
-        new_dataset = dataset.filter(self.process,
+        new_dataset = dataset.filter(self.process_batched,
                                      num_proc=self.runtime_np(),
                                      batch_size=self.batch_size,
                                      desc=self._name + '_process')