--- a/data_juicer/ops/filter/image_face_ratio_filter.py
+++ b/data_juicer/ops/filter/image_face_ratio_filter.py
@@ -74,7 +74,7 @@ class ImageFaceRatioFilter(Filter):
         self.model_key = prepare_model(model_type='opencv_classifier',
                                        model_path=cv_classifier)
 
-    def compute_stats(self, sample, context=False):
+    def compute_stats_single(self, sample, context=False):
         # check if it's computed already
         if StatsKeys.face_ratios in sample[Fields.stats]:
             return sample
@@ -112,7 +112,7 @@ class ImageFaceRatioFilter(Filter):
         ]
         return sample
 
-    def process(self, sample):
+    def process_single(self, sample):
         face_ratios = sample[Fields.stats][StatsKeys.face_ratios]
         if len(face_ratios) <= 0:
             return True