--- a/data_juicer/ops/filter/image_watermark_filter.py
+++ b/data_juicer/ops/filter/image_watermark_filter.py
@@ -58,7 +58,7 @@ class ImageWatermarkFilter(Filter):
             pretrained_model_name_or_path=hf_watermark_model,
             trust_remote_code=trust_remote_code)
 
-    def compute_stats(self, sample, rank=None, context=False):
+    def compute_stats_single(self, sample, rank=None, context=False):
         # check if it's computed already
         if StatsKeys.image_watermark_prob in sample[Fields.stats]:
             return sample
@@ -88,7 +88,7 @@ class ImageWatermarkFilter(Filter):
 
         return sample
 
-    def process(self, sample, rank=None):
+    def process_single(self, sample, rank=None):
         itm_probs = sample[Fields.stats][StatsKeys.image_watermark_prob]
         if len(itm_probs) <= 0:
             return True