- Extract the batched processing loop to a new function (e.g. compute_stats_batched), and decide the function branch to be called when processing data.
- Pros:
  - No need to modify each OP to support batched processing. Only need to set the arg `_batched_op` to True for each OP.
  - Developers can add new OPs in a single-sample-dict way, which is more friendly.
- Cons (for now):
  - Slightly slower data processing: 4 scenarios were tested and the results are shown below:

| Scenario                    | OPs                                      | Before   | After             |
|-----------------------------|------------------------------------------|----------|-------------------|
| Simple Filter               | `text_length_filter`                       | 4.850s   | 5.340s (+10.10%)  |
| Simple Mapper               | `whitespace_normalization_mapper`          | 12.145s  | 12.575s (+3.54%)  |
| With `context` (OP fusion)    | `word_num_filter` & `word_repetition_filter` | 80.456s  | 87.315s (+8.53%)  |
| With `rank` (Model-based OPs) | `image_text_similarity_filter`             | 562.586s | 590.502s (+4.96%) |